third_party/blink/web_tests/http/tests/navigation/pushstate-in-blob-denied.html

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function () {
    var blobUrl = URL.createObjectURL(new Blob());
    assert_throws_dom('SecurityError', function () {
        history.pushState(null, null, blobUrl);
    });
}, 'pushState to blob:-URL should fail with SecurityError');
</script>
